+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [v2.3.0](https://github.com/pojozhang/sugar/tree/v2.3.0)
+### Added
+- New `New` function to build a client.
+
+### Changed
+- Remove `NewClient` function.
+
 ## [v2.2.0](https://github.com/pojozhang/sugar/tree/v2.2.0)
 ### Added
 - New `Retryer` plugin.